decomp: multi-unit sweep — TrainerGemTab erase+int-loop, GemTrackDir, SaveLoadManager, Rnd, Utl, bandobj, world
Key wins:
- TrainerGemTab::DrawExtraTails 96.2→100% (int i + unsigned cast for cmplw, erase vs resize)
- TrainerGemTab::Draw 90.4→93.7% (erase(begin,end) removes spurious _M_fill_insert)
- GemTrackDir::PrepareChordMesh 91.2→96.5% (branch polarity + GeomOwner removal)
- SaveLoadManager::SetState 87.9→93.0% (erase replaces resize, case reordering)
- PatchDir::HasSticker 0→100% (Str()!=gNullStr + DECOMP_FORCEBLOCK)
- Rnd::RemovePointTest 100%, DrawTimers 90.3%
- Utl::ConvertBonesToTranses 98.9% (fp_contract off)
- Various bandobj/beatmatch/world/utl near-100% fixes from sweep

+++ src/system/bandobj/CharKeyHandMidi.cpp
@@ -53,7 +53,7 @@ CharIKFingers::FingerNum CharKeyHandMidi::FindPreferredFinger(
     KeyboardKey setToKey, KeyboardKey lastKeyDown, CharIKFingers::FingerNum lastFingerDown
 ) {
     int finger;
-    if (setToKey == lastKeyDown)
+    if ((int)(int)setToKey == lastKeyDown)
         return lastFingerDown;
     if (lastKeyDown == 0 || lastFingerDown == CharIKFingers::kFingerNone)
         return CharIKFingers::kFingerMiddle;
